Versión 1.0.4, #526: Corrección en la BD para soportar particiones GPT.
git-svn-id: https://opengnsys.es/svn/branches/version1.0@3043 a21b9725-9963-47de-94b9-378ad31fedc9

+++ admin/Database/ogAdmBD.sql
@@ -490,7 +490,7 @@ CREATE TABLE IF NOT EXISTS `ordenadores` (
 CREATE TABLE IF NOT EXISTS `ordenadores_particiones` (
   `idordenador` int(11) NOT NULL,
   `numpar` tinyint(4) NOT NULL,
-  `codpar` smallint(11) NOT NULL,
+  `codpar` int(8) NOT NULL,
   `tamano` int(11) NOT NULL,
   `idsistemafichero` smallint(11) NOT NULL,
   `idnombreso` smallint(11) NOT NULL,
